betting-analyzer/backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
apscheduler==3.10.4
httpx==0.28.1
curl_cffi>=0.7.0
supabase==2.11.0
python-dotenv==1.0.1
xgboost==2.1.3; python_version < "3.13"
scikit-learn==1.6.1; python_version < "3.13"
numpy==2.2.2; python_version < "3.13"
pandas==2.2.3; python_version < "3.13"
